+++ heypublisher-sub-mgr.php
@@ -2,10 +2,10 @@
 /*
 Plugin Name: HeyPublisher Submission Manager
 Plugin URI: https://github.com/HeyPublisher/heypublisher-submission-manager
-Description: HeyPublisher is a better way of managing unsolicited submissions directly within WordPress.
+Description: HeyPublisher is a better way of managing submissions for your WordPress-powered publication.
 Author: HeyPublisher
 Author URI: https://www.heypublisher.com
-Version: 3.0.1
+Version: 3.1.0
 Requires at least: 4.0
 
 
@@ -83,6 +83,8 @@ define('HEY_DIR', dirname(plugin_basename(__FILE__)));
   2.9.0 => 76
   3.0.0 => 80
   3.0.1 => 81
+  3.1.0 => 82
+
 
 ---------------------------------------------------------------------------------
 */
@@ -90,11 +92,11 @@ define('HEY_DIR', dirname(plugin_basename(__FILE__)));
 // Configs specific to the plugin
 // Build Number (must be a integer)
 define('HEY_BASE_URL', get_option('siteurl').'/wp-content/plugins/'.HEY_DIR.'/');
-define("HEYPUB_PLUGIN_BUILD_DATE", "2020-06-27");
+define("HEYPUB_PLUGIN_BUILD_DATE", "2020-08-30");
 // Version Number (can be text)
-define("HEYPUB_PLUGIN_BUILD_NUMBER", "81");  // This controls whether or not we get upgrade prompt
-define("HEYPUB_PLUGIN_VERSION", "3.0.1");
-define("HEYPUB_PLUGIN_TESTED", "5.3.0");
+define("HEYPUB_PLUGIN_BUILD_NUMBER", "82");  // This controls whether or not we get upgrade prompt
+define("HEYPUB_PLUGIN_VERSION", "3.1.0");
+define("HEYPUB_PLUGIN_TESTED", "5.5.0");
 
 # Base domain
 $domain = 'https://www.heypublisher.com';
@@ -112,14 +114,11 @@ define('HEYPUB_PLUGIN_FULLPATH', dirname(__FILE__));
 define('HEYPUB_FEEDBACK_EMAIL_VALUE','support@heypublisher.com?subject=HeyPublisher%20Wordpress%20Plugin');
 define('HEYPUB_FEEDBACK_GETSATISFACTION','mailto:support@heypublisher.com?Subject=HeyPublisher%20Submission%20Manager');
 // define('HEYPUB_SVC_URL_STYLE_GUIDE','https://blog.heypublisher.com/docs/plugins/wordpress/style_guide/');     # designates the URL of the style guide
-# designates the base URL and version of API
-define('HEYPUB_SVC_URL_BASE', $domain . '/api/v1');
-define('HEYPUB_API', $domain . '/api');
+define('HEYPUB_DOMAIN', $domain);
 # Stylesheet for plugin resides on HP server now
-define('HEYPUB_SVC_STYLESHEET_URL',$domain . '/stylesheets/wordpress/plugin.css?' . HEYPUB_PLUGIN_VERSION);
+define('HEYPUB_SVC_STYLESHEET_URL',HEYPUB_DOMAIN . '/stylesheets/wordpress/plugin.css?' . HEYPUB_PLUGIN_VERSION);
 
 // TODO: remove these defines and reference directly from XML code that makes query
-define('HEYPUB_SVC_URL_SUBMIT_FORM','submissions');
 define('HEYPUB_SVC_URL_AUTHENTICATE','publishers/fetch_or_create');           # initial plugin authentication
 define('HEYPUB_SVC_URL_GET_PUBLISHER','publishers/show');                     # update the options
 define('HEYPUB_SVC_URL_UPDATE_PUBLISHER','publishers/update_publisher');      # update the options
@@ -160,9 +159,11 @@ define('HEYPUB_POST_META_KEY_SUB_ID','_heypub_post_meta_key_sub_id');
 global $hp_xml, $hp_base, $hp_config;
 global $hp_updater;
 
-include_once(HEYPUB_PLUGIN_FULLPATH . '/updater.php');
-$hp_updater = new HeyPublisherUpdater( __FILE__ ); // instantiate our class
-$hp_updater->set_username( 'HeyPublisher' ); // set username
+if (!class_exists("\HeyPublisher\Base\Updater")) {
+  require_once(HEYPUB_PLUGIN_FULLPATH . '/include/classes/HeyPublisher/Base/Updater.class.php');
+}
+// initialize the updater and test for update
+$hp_updater = new \HeyPublisher\Base\Updater( __FILE__ );
 $hp_updater->set_repository( 'heypublisher-submission-manager' ); // set repo
 $hp_updater->initialize(HEYPUB_PLUGIN_TESTED); // initialize the updater
 
@@ -191,7 +192,6 @@ require_once(HEYPUB_PLUGIN_FULLPATH . '/admin/heypub-main.php');
 // Plugin configuration options
 load_template(HEYPUB_PLUGIN_FULLPATH . '/include/classes/HeyPublisher/Page/Options.class.php');
 $hp_opt = new \HeyPublisher\Page\Options;
-$hp_opt->domain = $domain;
 
 // Submissions
 load_template(HEYPUB_PLUGIN_FULLPATH . '/include/classes/HeyPublisher/Page/Submissions.class.php');
@@ -432,7 +432,7 @@ function heypub_uninit() {
     $opts['genres_list'] = false;
     $opts['guide'] = get_permalink($opts['sub_guide_id']);
     // TODO: This should change to api->update_publisher();
-    $hp_opt->api->deactivate();
+    $hp_opt->pubapi->deactivate();
   }
   $hp_config->install = false;
   $hp_config->config = false;