+++ include/classes/HeyPublisher/API/Email.class.php
@@ -6,21 +6,27 @@ if(preg_match('#' . basename(__FILE__) . '#', $_SERVER['PHP_SELF'])) { die('HeyP
 * Submission class for JSON API calls related to Submissions
 *
 */
+if (!class_exists("\HeyPublisher\Base\API")) {
+  require_once(HEYPUB_PLUGIN_FULLPATH . '/include/classes/HeyPublisher/Base/API.class.php');
+}
 
-require_once(HEYPUB_PLUGIN_FULLPATH . '/include/classes/HeyPublisher/API.class.php');
-class Email extends \HeyPublisher\API {
+class Email {
+
+  var $api = null;
+  var $logger = null;
 
   public function __construct() {
-  	parent::__construct();
+    global $HEYPUB_API, $HEYPUB_LOGGER;
+    $this->api = $HEYPUB_API;
+    $this->logger = $HEYPUB_LOGGER;
   }
 
   public function __destruct() {
-  	parent::__destruct();
   }
 
   public function get_emails() {
     $path = 'v2/email_templates';
-    $result = $this->get($path);
+    $result = $this->api->get($path);
     $this->logger->debug(sprintf("get_emails():\n\tResults: %s",print_r($result,1)));
     $emails = $this->normalize_results($result,true);
     return $emails;
@@ -28,14 +34,14 @@ class Email extends \HeyPublisher\API {
 
   public function get_email($id) {
     $path = sprintf('v2/email_templates/%s',$id);
-    $result = $this->get($path);
+    $result = $this->api->get($path);
     $emails = $this->normalize_results($result);
     return $emails[0];
   }
 
   public function get_submission_states() {
     $path = 'v2/email_templates/submission_states';
-    $result = $this->get($path);
+    $result = $this->api->get($path);
     $this->logger->debug(sprintf("get_submission_states():\n\tResults: %s",print_r($result,1)));
     if ($result['submission_states'] && sizeof($result['submission_states']) > 0 ) {
       return $result['submission_states'];
@@ -48,7 +54,7 @@ class Email extends \HeyPublisher\API {
   public function delete_template($id) {
     $message = 'Unknown DELETE error';
     $path = sprintf('v2/email_templates/%s',$id);
-    $result = $this->delete($path);
+    $result = $this->api->delete($path);
     if ($result == 'deleted') {
       $message = 'Email template deleted';
     }
@@ -63,7 +69,7 @@ class Email extends \HeyPublisher\API {
     $message = 'Unknown POST error';
     if ($post['hp_email']) {
       $path = 'v2/email_templates';
-      $result = $this->post($path,$post['hp_email']);
+      $result = $this->api->post($path,$post['hp_email']);
       $emails = $this->normalize_results($result);
       if ($emails[0] && $emails[0]['id'] > 0) {
         $message = 'Email template updated';